requirements.txt
BeautifulSoup==3.2.1
Django==1.10.5
Flask==0.12
Flask-SQLAlchemy==2.1
Jinja2==2.8
Markdown==2.6.7
MarkupSafe==0.23
Pillow==2.5.1
PyYAML==3.11
Pygments==2.1.3
SQLAlchemy==1.0.15
Werkzeug==0.11.15
altgraph==0.10.2
appnope==0.1.0
backports-abc==0.4
backports.shutil-get-terminal-size==1.0.0
bdist-mpkg==0.5.0
beautifulsoup4==4.3.2
bonjour-py==0.3
certifi==2016.9.26
chardet==2.3.0
click==6.7
configparser==3.5.0
cssselect==0.9.1
cvxopt==1.1.8
decorator==4.0.10
entrypoints==0.2.2
enum34==1.1.6
feedfinder2==0.0.1
feedparser==5.1.3
functools32==3.2.3-2
goose-extractor==1.0.25
html5lib==0.999999999
ipykernel==4.5.0
ipython==5.1.0
ipython-genutils==0.1.0
ipython-sql==0.3.6
ipywidgets==5.2.2
itsdangerous==0.24
jieba==0.35
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.4.0
jupyter-console==5.0.0
jupyter-core==4.2.0
lxml==3.3.5
macholib==1.5.1
matplotlib==1.3.1
mistune==0.7.3
modulegraph==0.10.4
nbconvert==4.2.0
nbformat==4.1.0
newspaper==0.0.9.8
nltk==2.0.5
notebook==4.2.3
numpy==1.11.2
olefile==0.44
pandas==0.18.1
pathlib2==2.1.0
pexpect==4.2.1
pickleshare==0.7.4
prettytable==0.7.2
prompt-toolkit==1.0.7
ptyprocess==0.5.1
py2app==0.7.3
pyOpenSSL==0.13.1
pyobjc-core==2.5.1
pyobjc-framework-Accounts==2.5.1
pyobjc-framework-AddressBook==2.5.1
pyobjc-framework-AppleScriptKit==2.5.1
pyobjc-framework-AppleScriptObjC==2.5.1
pyobjc-framework-Automator==2.5.1
pyobjc-framework-CFNetwork==2.5.1
pyobjc-framework-Cocoa==2.5.1
pyobjc-framework-Collaboration==2.5.1
pyobjc-framework-CoreData==2.5.1
pyobjc-framework-CoreLocation==2.5.1
pyobjc-framework-CoreText==2.5.1
pyobjc-framework-DictionaryServices==2.5.1
pyobjc-framework-EventKit==2.5.1
pyobjc-framework-ExceptionHandling==2.5.1
pyobjc-framework-FSEvents==2.5.1
pyobjc-framework-InputMethodKit==2.5.1
pyobjc-framework-InstallerPlugins==2.5.1
pyobjc-framework-InstantMessage==2.5.1
pyobjc-framework-LatentSemanticMapping==2.5.1
pyobjc-framework-LaunchServices==2.5.1
pyobjc-framework-Message==2.5.1
pyobjc-framework-OpenDirectory==2.5.1
pyobjc-framework-PreferencePanes==2.5.1
pyobjc-framework-PubSub==2.5.1
pyobjc-framework-QTKit==2.5.1
pyobjc-framework-Quartz==2.5.1
pyobjc-framework-ScreenSaver==2.5.1
pyobjc-framework-ScriptingBridge==2.5.1
pyobjc-framework-SearchKit==2.5.1
pyobjc-framework-ServiceManagement==2.5.1
pyobjc-framework-Social==2.5.1
pyobjc-framework-SyncServices==2.5.1
pyobjc-framework-SystemConfiguration==2.5.1
pyobjc-framework-WebKit==2.5.1
pyparsing==2.0.1
pyquery==1.2.17
python-dateutil==2.4.0
pytz==2013.7
pyzmq==15.4.0
qtconsole==4.2.1
requests==2.3.0
scipy==0.18.1
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
sqlparse==0.2.1
tabulate==0.7.5
terminado==0.6
tldextract==1.5.1
tornado==4.4.1
traitlets==4.3.0
vboxapi==1.0
virtualenv==13.1.2
wcwidth==0.1.7
webencodings==0.5
wget==3.2
widgetsnbextension==1.2.6
wsgiref==0.1.2
xattr==0.6.4
zope.interface==4.1.1
